% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E-formatTSTable.R
\name{formatTSTable}
\alias{formatTSTable}
\title{Formats output from computeTimeSeriesStatistics for use in reproducible research output}
\usage{
formatTSTable(TSTable, benchmark = NA)
}
\arguments{
\item{TSTable}{an xts vector or timeseries matrix OR the resulting data.frame output from computeTimeSeriesStatistics}

\item{benchmark}{an optional benchmark for use with an xts TSTable object (ignored if TSTable is a data.frame)}
}
\value{
a formatted kable table
}
\description{
Formats output from computeTimeSeriesStatistics for use in reproducible research output
}
\examples{
none


}

